backend/requirements.txt
# FastAPI Core
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# AI - Google Gemini
google-generativeai==0.3.2

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1

# Redis Cache
redis==5.0.1
hiredis==2.3.2

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# XML Processing
lxml==5.1.0
xmltodict==0.13.0

# Code Execution & Sandboxing
docker==7.0.0

# Utilities
python-dotenv==1.0.0
httpx==0.26.0
aiofiles==23.2.1

# Monitoring & Logging
loguru==0.7.2

# Testing (Development)
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0
